Rename Canvas/FormBoard to FormPreview
