+++ frontend/form-builder/src/app/App.module.css
@@ -19,7 +19,7 @@ body {
   display: grid;
   grid-template-columns: 1fr 6fr 3fr;
   gap: var(--fds-spacing-2);
-  min-height: 600px;
+  height: 100%;
 }
 
 .builderSection {
@@ -28,4 +28,10 @@ body {
   border-radius: var(--fds-border_radius-medium);
   padding: var(--fds-spacing-5);
   text-align: center;
+  margin: auto 0;
+  height: 100%;
+  
+  /* 100vh - header - subheader and margins - footer */
+  /* disables global scrollbar */
+  max-height: calc(100vh - var(--fds-sizing-26) - calc(3 * var(--fds-sizing-10)) - calc(2 * var(--fds-spacing-30)));
 }
