--- frontend/form-builder/src/components/FormBoard/index.ts
@@ -1,1 +1,0 @@
-export { FormBoard } from "./FormBoard";
