+++ frontend/form-builder/src/components/FormPreview/FormPreview.module.css
@@ -1,10 +1,18 @@
-.noComponents {
+.previewHeading {
+  margin-bottom: var(--fds-spacing-4);
+}
+
+.dropArea {
   /* Center the 'drag components here' text in both axis */
-  height: 80%;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  height: 90%;
+  overflow-y: auto;
+  overflow-x: hidden;
+  scrollbar-width: thin;
+  scrollbar-color: #2c3e50 transparent;
 }
 
 .noComponentsIcon {
