+++ frontend/form-builder/src/components/FormPreview/index.ts
@@ -1,0 +1,1 @@
+export { FormPreview } from "./FormPreview";
